Further improvements to class structure:
Changed the class structure.

The class structure was meaningless before because Object3D class was only rendering a Utah Teapot.
Instead of this, I changed Object3D class to an abstract class and made a UtahTeapot class that
inherits Object3D. It is still open to improvements and more abstractions can be made.